backend/requirements.txt
Flask==2.3.3
tensorflow==2.13.0
keras==2.13.1
opencv-python==4.8.0.76
pandas==2.0.3
numpy==1.24.3
scikit-learn==1.3.0
matplotlib==3.7.2
seaborn==0.12.2
plotly==5.15.0
streamlit==1.25.0
Pillow==10.0.0
requests==2.31.0
flask-cors==4.0.0
gunicorn==21.2.0